+++ Shop.Domain/Entities/CartItem.cs
@@ -10,7 +10,7 @@ namespace Shop.Domain.Entities
     {
         public int Id { get; set; }
         public int Quantity { get; set; }
-        public float UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public Cart Cart { get; set; }
         public string CartId { get; set; }
         public Item Item { get; set; }
